refactor: Move single item parsers to their own module
